fix(access): finish lifecycle cancellation boundaries
